c++ version with error fmt undefined
